feat(video): vector2d position pad in video modal; text position + blur in video-text
Extract the draggable 2D pad out of ControlledVector2DInput into a
form-agnostic Vector2DPad (value/onChange), so it can be reused outside the
react-hook-form renderer. ControlledVector2DInput becomes a thin wrapper that
preserves sibling keys.

- Video asset modal: replace the separate Position X / Position Y sliders with
  a single Vector2DPad (centered [-1, 1], y-down to match computeVideoLayout).
- video-text: add a "Text position" vector2d field that offsets the masking
  text (0.5/0.5 stays centered), and a "Video blur" slider that softens the
  composited video layer while keeping the text mask crisp.
