+++ src/components/ClientProcessingSketch/components/TemplateOptions/components/ContentItems/components/ControlledVector2DInput/ControlledVector2DInput.tsx
@@ -1,85 +1,29 @@
 "use client";
 
 import {
-  useRef
-} from "react";
-import {
   useController, useFormContext
 } from "react-hook-form";
 
-import clamp from "@/utils/clamp";
-import {
-  fractionToValue,
-  stepDecimals,
-  valueToFraction,
-  type AxisRange
-} from "./utils/vector2dMath";
+import Vector2DPad, {
+  type Vector2DInputConfig, type Vector2DValue
+} from "./Vector2DPad";
 
-/**
- * Subset of the field config understood by the pad. Kept independent from
- * `FieldConfig` so the component stays reusable outside the form renderer.
- */
-export interface Vector2DInputConfig {
-  /**
-   * When false, both axes are constrained to non-negative values ([0, max]) so
-   * the vector can only point up/right — handy for scaling a vector's strength
-   * down to 0 without flipping its direction. Defaults to true, which yields a
-   * centered pad spanning [min, max] on both axes.
-   */
-  allowNegative?: boolean;
-  /** Shared lower bound. Defaults to -1 (or 0 when `allowNegative` is false). */
-  min?: number;
-  /** Shared upper bound. Defaults to 1. */
-  max?: number;
-  /** Shared snapping increment. Defaults to 0.01. */
-  step?: number;
-  /** Per-axis overrides, merged over the shared min/max/step. */
-  xAxis?: Partial<AxisRange>;
-  yAxis?: Partial<AxisRange>;
-  /**
-   * Invert the vertical axis so the top of the pad maps to the *minimum* value.
-   * Use it for screen-space positions (where y grows downward): dragging the
-   * handle up then moves the point toward the top of the canvas. Defaults to
-   * false (top = max), matching the "Y points up" convention used for vectors.
-   */
-  yDown?: boolean;
-}
-
-type Vector2DValue = {
-  x: number;
-  y: number;
-};
+export type {
+  Vector2DInputConfig, Vector2DValue
+} from "./Vector2DPad";
 
 type Props = {
   name: string;
   config?: Vector2DInputConfig;
 };
 
-function resolveAxes( config: Vector2DInputConfig ): {
-  xAxis: AxisRange;
-  yAxis: AxisRange;
-} {
-  const allowNegative = config.allowNegative ?? true;
-  const min = config.min ?? ( allowNegative ? -1 : 0 );
-  const max = config.max ?? 1;
-  const step = config.step ?? 0.01;
-
-  return {
-    xAxis: {
-      min: config.xAxis?.min ?? min,
-      max: config.xAxis?.max ?? max,
-      step: config.xAxis?.step ?? step
-    },
-    yAxis: {
-      min: config.yAxis?.min ?? min,
-      max: config.yAxis?.max ?? max,
-      step: config.yAxis?.step ?? step
-    }
-  };
-}
-
+/**
+ * Binds the presentational {@link Vector2DPad} to react-hook-form. Reads/writes
+ * `{ x, y }` on the field while preserving any sibling keys stored alongside
+ * them, so the pad can edit a position nested inside a larger value object.
+ */
 export default function ControlledVector2DInput( {
-  name, config = {}
+  name, config
 }: Props ) {
   const {
     control
@@ -92,286 +36,20 @@ export default function ControlledVector2DInput( {
     control
   } );
 
-  const padRef = useRef<HTMLDivElement>( null );
-
-  const {
-    xAxis, yAxis
-  } = resolveAxes( config );
-
-  const yDown = config.yDown ?? false;
-
-  // Convert between a Y value and its vertical position in the pad (a 0..1
-  // fraction where 0 is the top). Unless yDown is set, the axis is flipped so
-  // the top of the pad represents the maximum value.
-  const valueYToFraction = ( y: number ) => {
-    const fraction = valueToFraction(
-      y,
-      yAxis
-    );
-
-    return yDown ? fraction : 1 - fraction;
-  };
-
-  const fractionToValueY = ( fraction: number ) => fractionToValue(
-    yDown ? fraction : 1 - fraction,
-    yAxis
-  );
-
   const raw = field.value as Partial<Vector2DValue> | undefined;
-  const value: Vector2DValue = {
-    x: typeof raw?.x === "number" ? raw.x : xAxis.min,
-    y: typeof raw?.y === "number" ? raw.y : yAxis.min
-  };
-
-  // Preserve any sibling keys on the stored object while updating x/y.
-  const commit = ( next: Vector2DValue ) => {
-    field.onChange( {
-      ...( raw && typeof raw === "object" ? raw : {} ),
-      x: next.x,
-      y: next.y
-    } );
-  };
-
-  const updateFromPointer = (
-    clientX: number, clientY: number
-  ) => {
-    const pad = padRef.current;
-
-    if ( !pad ) {
-      return;
-    }
-
-    const rect = pad.getBoundingClientRect();
-    const fractionX = ( clientX - rect.left ) / rect.width;
-    const fractionY = ( clientY - rect.top ) / rect.height;
-
-    commit( {
-      x: fractionToValue(
-        fractionX,
-        xAxis
-      ),
-      y: fractionToValueY( fractionY )
-    } );
-  };
-
-  const handlePointerDown = ( event: React.PointerEvent<HTMLDivElement> ) => {
-    event.currentTarget.setPointerCapture( event.pointerId );
-    updateFromPointer(
-      event.clientX,
-      event.clientY
-    );
-  };
-
-  const handlePointerMove = ( event: React.PointerEvent<HTMLDivElement> ) => {
-    if ( !event.currentTarget.hasPointerCapture( event.pointerId ) ) {
-      return;
-    }
-
-    updateFromPointer(
-      event.clientX,
-      event.clientY
-    );
-  };
-
-  const nudge = (
-    dx: number, dy: number
-  ) => {
-    const round = (
-      v: number, step: number | undefined
-    ) => Number( v.toFixed( stepDecimals( step ) ) );
-
-    // `dy` is expressed in screen terms (+1 = up). Flip it for a non-inverted
-    // axis so pressing "up" increases the value.
-    const valueDy = yDown ? -dy : dy;
-
-    commit( {
-      x: clamp(
-        round(
-          value.x + dx * ( xAxis.step ?? 0.01 ),
-          xAxis.step
-        ),
-        xAxis.min,
-        xAxis.max
-      ),
-      y: clamp(
-        round(
-          value.y + valueDy * ( yAxis.step ?? 0.01 ),
-          yAxis.step
-        ),
-        yAxis.min,
-        yAxis.max
-      )
-    } );
-  };
-
-  const handleKeyDown = ( event: React.KeyboardEvent<HTMLDivElement> ) => {
-    const moves: Record<string, [number, number]> = {
-      ArrowLeft: [
-        -1,
-        0
-      ],
-      ArrowRight: [
-        1,
-        0
-      ],
-      ArrowUp: [
-        0,
-        1
-      ],
-      ArrowDown: [
-        0,
-        -1
-      ]
-    };
-
-    const move = moves[ event.key ];
-
-    if ( move ) {
-      event.preventDefault();
-      nudge(
-        move[ 0 ],
-        move[ 1 ]
-      );
-    }
-  };
-
-  const handleAxisInput = (
-    axis: "x" | "y", rawValue: string
-  ) => {
-    const parsed = parseFloat( rawValue );
-
-    if ( Number.isNaN( parsed ) ) {
-      return;
-    }
-
-    const range = axis === "x" ? xAxis : yAxis;
-    const clamped = clamp(
-      parsed,
-      range.min,
-      range.max
-    );
-
-    commit( axis === "x"
-      ? {
-        ...value,
-        x: clamped
-      }
-      : {
-        ...value,
-        y: clamped
-      } );
-  };
-
-  // Handle position as a percentage within the pad (top-left origin). The
-  // guide line always emanates from the geometric centre so the pad reads
-  // like a small graph, whatever the axis ranges.
-  const pointLeft = valueToFraction(
-    value.x,
-    xAxis
-  ) * 100;
-  const pointTop = valueYToFraction( value.y ) * 100;
-
-  const numberInputClassName =
-    "w-full text-center text-xs font-mono px-1 py-0.5 rounded border border-theme/30 bg-theme/20 focus:outline-none focus:ring-1 focus:ring-theme";
 
   return (
-    <div className="flex w-full max-w-[100px] flex-col gap-1">
-      <div className="flex items-center gap-1">
-        <input
-          type="number"
-          aria-label={ `${ name } x` }
-          className={ numberInputClassName }
-          value={ value.x }
-          min={ xAxis.min }
-          max={ xAxis.max }
-          step={ xAxis.step }
-          onChange={ ( event ) => handleAxisInput(
-            "x",
-            event.target.value
-          ) }
-        />
-        <input
-          type="number"
-          aria-label={ `${ name } y` }
-          className={ numberInputClassName }
-          value={ value.y }
-          min={ yAxis.min }
-          max={ yAxis.max }
-          step={ yAxis.step }
-          onChange={ ( event ) => handleAxisInput(
-            "y",
-            event.target.value
-          ) }
-        />
-      </div>
-
-      <div
-        ref={ padRef }
-        role="application"
-        tabIndex={ 0 }
-        aria-label="2D vector pad"
-        onPointerDown={ handlePointerDown }
-        onPointerMove={ handlePointerMove }
-        onKeyDown={ handleKeyDown }
-        className="relative aspect-square w-full cursor-crosshair touch-none select-none overflow-hidden rounded-lg border border-theme bg-background/50 focus:outline-none focus:ring-1 focus:ring-theme"
-      >
-        <svg
-          className="absolute inset-0 h-full w-full"
-          viewBox="0 0 100 100"
-          preserveAspectRatio="none"
-          aria-hidden="true"
-        >
-          {/* Static grid through the geometric centre. */}
-          <line
-            x1="50"
-            y1="0"
-            x2="50"
-            y2="100"
-            className="stroke-theme/40"
-            strokeWidth={ 0.5 }
-            strokeDasharray="2 2"
-            vectorEffect="non-scaling-stroke"
-          />
-          <line
-            x1="0"
-            y1="50"
-            x2="100"
-            y2="50"
-            className="stroke-theme/40"
-            strokeWidth={ 0.5 }
-            strokeDasharray="2 2"
-            vectorEffect="non-scaling-stroke"
-          />
-
-          {/* Guide line from the centre of the pad to the current value. */}
-          <line
-            x1={ 50 }
-            y1={ 50 }
-            x2={ pointLeft }
-            y2={ pointTop }
-            className="stroke-foreground/60"
-            strokeWidth={ 1 }
-            strokeDasharray="2 2"
-            vectorEffect="non-scaling-stroke"
-          />
-        </svg>
-
-        {/* Tiny axis labels so the active axis stays obvious. */}
-        <span className="pointer-events-none absolute right-0.5 top-1/2 -translate-y-1/2 bg-background/70 px-px font-mono text-[7px] leading-none text-gray-400">
-          x
-        </span>
-        <span className="pointer-events-none absolute left-1/2 top-0.5 -translate-x-1/2 bg-background/70 px-px font-mono text-[7px] leading-none text-gray-400">
-          y
-        </span>
-
-        <span
-          className="pointer-events-none absolute h-2.5 w-2.5 -translate-x-1/2 -translate-y-1/2 rounded-full border border-background bg-foreground shadow"
-          style={ {
-            left: `${ pointLeft }%`,
-            top: `${ pointTop }%`
-          } }
-        />
-      </div>
-    </div>
+    <Vector2DPad
+      value={ raw }
+      onChange={ ( next ) =>
+        field.onChange( {
+          ...( raw && typeof raw === "object" ? raw : {} ),
+          x: next.x,
+          y: next.y
+        } )
+      }
+      config={ config }
+      ariaLabel={ name }
+    />
   );
 }
